webhook-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
httpx==0.25.2
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1